Move declarative helpers to opengever.base.model.
